+++ react/Components/collapsibleArea.tsx
@@ -1,8 +1,10 @@
 import React, { FC } from 'react'
+import { useIntl } from 'react-intl'
 import { Collapsible } from 'vtex.styleguide'
 
 import { useStore } from '../hooks/useStore'
 import '../styles.global.css'
+import { collapsible } from '../utils/definedMessages'
 import AlertArea from './alertArea'
 import InputButtonArea from './inputButtonArea'
 import TitlesArea from './titlesArea'
@@ -10,10 +12,17 @@ import TitlesArea from './titlesArea'
 const CollapsibleArea: FC = () => {
   const { setButton, button } = useStore()
 
+  const intl = useIntl()
+
   return (
     <div className="w-80 center">
       <div className="mb5 mt5">
         <Collapsible
+          header={
+            <h1 className="t-heading-2">
+              {intl.formatMessage(collapsible.title)}
+            </h1>
+          }
           onClick={() => setButton(!button)}
           isOpen={button}
           align="right"